Don't annotate rendered view with filenames
While the feature looks useful, it isn't compatible with the way we
render the `Layout::CommonHtmlAttributesComponent` because we render it
inside an HTML tag.

Instead of editing the commit where we enabled this option (the one
upgrading to Rails 8.0), we're changing this option in a separate commit
so it's easier to know the reason why we don't use the
`annotate_rendered_view_with_filenames` feature.