Use a shared component to render basic info
We were using the same code in five different places.

Note we're adding a `management_polymorphic_path` which is quite a hack
as it makes an exception for budget investments. The reason is that in
admin and valuation routes we define:

```
resources :budgets
  # (...)
  resources :budget_investments
```

While in management routes we define:

```
resources :budgets
  # (...)
  resources :investments
```

Furthermore, in the admin section, we use the
`Admin::BudgetInvestmentsController`, while in the management section we
use the `Management::Budgets::InvestmentsController`.

Long story short: we can't use the same code to calculate a management
polymorphic route and an admin polymorphic route for budget investments.
So we're directly calling the `management_budget_investment_path` method
instead in `management_polymorphic_path` when the resource is an
investment.

Also note we're replacing the `topic-info` HTML class with `basic-info`
in both dashboard comments info and community poll info. We're doing so
because they display very basic information, even if they're inside an
element with a `topic-show` class. This change slightly affects the
styles, but nothing significant; the margins are now a bit different.