Use CSS to style separators in basic info component
Using CSS instead of HTML entities is what we've been doing during the
last few years.

We're getting some false positives when checking for hard coded strings
with ERB Lint; now we've reduced the amount of reported hard coded
strings from 87 to 66.

Note we're using `::after` pseudoelements because, in the case of
comments, the `::before` element is used for the comments icon. Also
note we're using `position: absolute` for the separator; if we didn't,
in elements with a background like the "collective" tag for
organizations or official position labels, the separator would be inside
the area with a different background, which isn't what we want.