Extract methods to show percentages in poll stats/results
We were using the same code several times.

Note that, with these changes, ERB Lint no longer reports hard coded
strings because of the parenthesis we use for the percentages. ERB Lint
can't detect them because they're no longer in an ERB file However, the
problem still exists since some languages might not use parenthesis in
these cases. So we might have to revisit this issue in the future.