Use a shared component to render detailed information
We were using the same code in six different places.

Note we're changing the conditition to render the flag/unflag actions
from `if current_user` to `if can?(:flag, record)`. The latter is more
reliable, and it works with topics (that can't be flagged even if
there's a current user).