Fix typo in budget investment info CSS
Due to this undetected typo, the sytles defined here weren't applied to
the `.budget-investment-show .budget-investment-info` selector. To apply
these styles, as a workaround we were adding `.budget-investment-show`
to the list of selectors that are used in `index` actions. And then we
were also adding a `<br>` tag because this meant the bottom margin was
too small.

By fixing the typo, we can remove the workarounds.