Move flag actions partial to a component
Note that we're slightly changing the CSS. Now the `.flag-disable` and
`.flag-active` selector are nested inside `.flag-content`. We're also
removing the `flag-disable` HTML class when rendering a flag in
moderation content. The reason why we're doing so is that in moderation
content this flag isn't a button, so we aren't applying the same styles
that we apply when the flag is a button.